mcp-openfda-client/pyproject.toml
[project]
name = "mcp-openfda-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]
